malloc: rewrite with atomic support

+++ include/ardix/syscall.h
@@ -16,17 +16,17 @@ enum syscall {
 };
 
 /** Perform an indirect system call. */
-int syscall(enum syscall number, ...);
+long syscall(enum syscall number, ...);
 
 /** The table of system call handlers, indexed by syscall number. */
-extern int (*const sys_table[NSYSCALLS])(sysarg_t arg1, sysarg_t arg2, sysarg_t arg3,
-					 sysarg_t arg4, sysarg_t arg5, sysarg_t arg6);
+extern long (*const sys_table[NSYSCALLS])(sysarg_t arg1, sysarg_t arg2, sysarg_t arg3,
+					  sysarg_t arg4, sysarg_t arg5, sysarg_t arg6);
 
 /* catchall handler that returns -ENOSYS */
-int sys_stub(void);
+long sys_stub(void);
 
-int sys_read(int fd, void *buf, size_t len);
-int sys_write(int fd, const void *buf, size_t len);
+long sys_read(int fd, void *buf, size_t len);
+long sys_write(int fd, const void *buf, size_t len);
 
 /*
  * This file is part of Ardix.
